[v4v] Fix module refcnt.
V4VIOCACCEPT allocates a new struct file without bumping the refcnt for
the module. The struct file is assigned v4v_fops_stream and a fd is
returned to userspace. When this fd is closed, the release() callback in
v4v_fops_stream file operations is called and the kernel will decrement
the module refcnt.
To maintain refcnt coherency, it has to be incremented when the struct
file is created, so when the fds are closed by userland, the refcnt does
not get decremented twice where is was only incremented once. Anyway,
returning a new fd should increase the refcnt.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>