libv4v: override soname library name.
A lot of scripts assume libv4v-1.0.so.0 or even libv4v-1.0.so.0.0.0 to
exist and have hardcoded that value.
It makes little difference to maintain the soname for now since the
version is unlikely to change soon, but will most certainly break the
various scripts trying to use the interposer.

Defer that for another time.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-978