Mark v4v file descriptors read-write
With f_flags unset, it defaults to 0, which is O_RDONLY.  Applications
or libraries will that check the file descriptor flags with fcntl
F_GETFL will think it is non writable.

We mark all v4v file descriptors as read-write since they are.  This
matches the behaviour of the linux network stack to mark all fds O_RDWR.

OXT-1304

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>