argo-module: add new <linux/argo.h>, breaks dependency cycle
Replaces use of the <xen/argo.h> from the hypervisor by Linux userspace
and removes dependency on the xen-tools package. This breaks a
dependency cycle which allows for refactoring and simplification of the
Xen recipes.

The new <linux/argo.h> is defined to avoid mandatory dependency on
<xen/interface/xen.h> which would force introducing a dependency on the
staging kernel headers, since the libc kernel headers are not sufficient
-- they do not include the xen interface header -- which would then make
the recipe MACHINE-specific.

OXT-1694

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>