Fix crash from SMAP violation
This crash is due to a strange case relation to how argo_write sends data.
As argo has no need to access data it sends it does not call
copy_from_user on the buffer sent to the hypervisor. Thus SMAP prevents
argo from reading the buffer causing argo to crash when this debug call
is enabled. This crash does not appear in dom0 due to it's higher
user privilede. To enable this debug call a flag must be set using stac()
and unset using clac() after call. As this is only necesary to read sent
buffer these calls must be explicitly enable in the build.

Signed-off-by: Nick Krasnoff <krasnoffn@ainfosec.com>