argo: Switch argo_copy_out to a loop
There are three copy calls from the ring to the buffer.  Consolidate
them down into 1 copy inside a loop.  The loop will be run one or two
times. Either a single unwrapped copy, or a copy of the tail of the ring
and then the head of the ring.

This paves the way for using copy_to_user.

There is a subtlety in argo_copy_out that when buf is NULL, it discards
the ring contents.  This is used by argo_interrupt_rx passing a literal
NULL for RING_TYPE_IDLE, so we must support it.

When discarding some or part of a message, we need to update the ring
pointer for the entire messages.  i.e. the difference between the
copy-out buffer "t" and the message length "len".

Without correctly discarding the bytes, recover_ring ends up getting
called with "Something went horribly wrong in a ring - dumping and
attempting a recovery".  This was observed with NDVM and the UI was
showing "send: resource vanished (Broken pipe)" error messages.

Ring-wrap handling need to be adjusted.  With the discard_len addition,
rxp can become >= ring_size, so it needs to wrap around as appropriate.
The earlier rxp changes are okay to check ring_size.  The first because
rxp is message header-aligned and it is advancing by message header, and
the second because the while loop can only advance to ring_size as a
maximum.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>