argo: Use workqueue for interrupt
Locking in interrupt context is problematic since you cannot sleep.  The
codebase has many issues with locking and required many changes to
spinlock_irqsave to deal with them.

Interrupt time should be minimized, so move the processing into a
workqueue.  Now the interrupt just kicks off the workqueue, and the
workqueue runs in process context.  This doesn't do much now, but it
sets up further removal of atomic locks.

Threaded interrupts would be an option, but xen doesn't have code to
bind to such interrupts.

Since the work is iterating through the list of rings, we don't want to
lose the interrupt if it's for a ring that was already processed.
schedule_work will either:
Queue the work if it's not queued
Not queue the work if it's already queued
Queue the work if it's running

So letting the interrupt handler call schedule_work will ensure
argo_work_fn is called or an addition call is scehduled if it is already
running.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>