argo: Always free pending_recv_list
kmemleak shows some leaks for pending_recv stuctures.  Only the sponsor
of the ring was freeing its pending_recv_list, but testing shows the
non-sponsor argo_privates may have pending_recv_list entries.

Move the freeing later in the function and always run it.  argo_private
always initializes pending_recv_list, so this should be safe.  If
anything is there, it should be freed.

While doing this, grab the pending_recv_lock for safety.   The comment
in argo_private says we are supposed to hold it.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>