argo: Use kvmalloc for potentially large pending_recv
The pending_recv allocations are potentially quite large.  They are
responsible for the order 4 allocation failures sometimes seen.

kvmalloc is our solution.  It was added to the linux kernel 4.12.  For
small allocations and an initial attempt at large allocations, it does
the virtually & physically contiguous kmalloc.  If that fails, it falls
back to vmalloc which is only virtually contiguous.  This is fine for
our purposes - we only need virtually contiguous.

This patch seems to avoid allocation failures which were triggerable
without it.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>