argo: Check p->r early in argo_release
p->r is checked early in argo_release, but inconsistently afterward.
p->r is only set if argo_bind is called.  It isn't a problem
currently, since p->state will still be ARGO_STATE_IDLE.

Move the p->r check to the start of the function and bypass most of it
when NULL.  This will allow simplifying the do/while loop in a
subsequent cleanup.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>